updating readmes, some more re-organization, added base spherical roller script (needs some more updating in the analyze section)

+++ spherical roller/img_to_sphere.sh
@@ -8,8 +8,9 @@ yellow="$(tput setaf 3 )"
 display_help() {
     echo "usage: .\\$(basename "$0") [analyze] { --help | -v --version | ...options } image" >&2
     echo
-    echo "script that uses openscad to generate a texture roller from an image input."
+    echo "script that uses openscad to generate a spherical texture roller from an image input."
     echo "outputs an stl file for 3d printing"
+    echo "NOTE: input image needs to be an equirectangular projection with a 2:1 aspect ratio"
     echo
     echo -e "\t--help\t\tdisplay this message"
     echo -e "\t-v, --version\t\t display version"
@@ -17,58 +18,34 @@ display_help() {
     echo -e "\t-w, --width\t\tresize image to specified width before generating stl (images larger than 1024x1024 can take quite a while)"
     echo -e "\t-ppi, --ppi\t\tthe ppi the final print will have [ppi] (default 72, recommended values 72-300)"
     echo -e "\t-d, --depth\t\tthe depth of the grooves on the roller [mm] (default 5, recommended values: 3-8)"
-    echo -e "\t-t, --thickness\t\tthe thickness of the cylinder wall [mm] (default 5)"
     echo -e "\t-o, --output\t\toutput file (default ./roller.stl)"
     echo
     echo "analyze:"
-    echo "usage: .\\$(basename "$0") analyze { -h --height | -w --width | -ar --aspect-ratio | -ppi --ppi | -r --radius | -l --length | -t --thickness }"
+    echo "usage: .\\$(basename "$0") analyze { -h --height | -w --width | -ppi --ppi | -r --radius }"
     echo
     echo "specify known variables to find what the unkown variables should be"
     echo
-    echo -e "\t-ar, --aspect-ratio\t\taspect ratio of the input image (x:y format)"
     echo -e "\t-r, --radius\t\tinner radius of final roller cylinder [in]"
-    echo -e "\t-l, --length\t\tlength of final roller roller cyinder [in]"
     echo
     exit 1
 }
 display_version() {
     echo "$(basename "$0") 0.1"
     echo "copyright (c) 2026 michael turenne and contributors"
-    # TODO: update this url
     echo "report bugs at: https://github.com/disdyakis/ceramics-tools/issues"
 }
-# summary function (args order): width, height, ppi, thickness, depth
+# summary function (args order): width, height, ppi, depth
 summary() {
     echo -ne "\n"
     echo "image"
     echo -e "\tdimensions: ${1}x${2}"
-    echo -e "\taspect ratio: $(aspect-ratio "$1" "$2")"
+    echo -e "\taspect ratio: 2:1"
     echo "roller"
     echo -e "\tppi: $3"
-    echo -e "\tlength: $(bc <<<"scale=2;$2/$3")in."
-    echo -e "\tradius: $(bc -l <<<"scale=2;($1/((8*a(1))*$3))-(0.039*$4)")in."
-    echo -e "\trim thickness: ${4}mm"
-    if [ -n "$5" ]; then
-        echo -e "\ttexture depth: ${5}mm"
-    fi
+    echo -e "\tradius: $(bc -l <<<"scale=2;($1/((8*a(1))*$3))")in."
     echo
 }
 # helper functions
-gcd() (
-    ! (( $1 % $2 )) && echo "$2" || gcd "$2" $(( $1 % $2 ))
-)
-aspect-ratio() (
-    local d=$(gcd "$1" "$2")
-    if [[ $d -eq 1 ]]; then
-        if [[ $1 -gt $2 ]]; then
-            echo $(bc <<<"scale=2;$1/$2" | sed '/\./ s/\.\{0,1\}0\{1,\}$//'):$(bc <<<"scale=2;$2/$2" | sed '/\./ s/\.\{0,1\}0\{1,\}$//')
-        else
-            echo $(bc <<<"scale=2;$1/$1" | sed '/\./ s/\.\{0,1\}0\{1,\}$//'):$(bc <<<"scale=2;$2/$1" | sed '/\./ s/\.\{0,1\}0\{1,\}$//')
-        fi
-    else
-        echo $(($1/$d)):$(($2/$d))
-    fi
-)
 solved() (
     local count=0
     for i in "${which_solved[@]}"
@@ -83,7 +60,6 @@ solved() (
 # defaults
 ppi=72
 depth=8
-thickness=5
 dir=$(dirname "$0")
 output=roller.stl
 # parse args
@@ -95,7 +71,7 @@ case $1 in
     -v | --version)
         display_version "$0"
         ;;
-    -o | --output | -ppi | --ppi | -d | --depth | -t | --thickness | -h | --height | -w | --width | -l | --length | -ar | --aspect-ratio | -r | --radius | analyze) ;;
+    -o | --output | -ppi | --ppi | -d | --depth | -h | --height | -w | --width | -l | --length | -r | --radius | analyze) ;;
     *)
         reg="^[\-]+"
         if [[ $1 =~ $reg ]]; then
@@ -117,10 +93,8 @@ while [ $# -gt $((1-analyze)) ] ; do
         ppi_set=1
         ;;
     -d | --depth) depth=$2 ;;
-    -t | --thickness) thickness=$2 ;;
     -h | --height) h=$2 ;;
     -w | --width) w=$2 ;;
-    -l | --length) l=$2 ;;
     -r | --radius) r=$2 ;;
     -ar | --aspect-ratio) ar=$2 ;;
     analyze)
@@ -143,38 +117,31 @@ if [[ $analyze -eq 1 ]]; then
         unset ppi
     fi
     # array slots = (height width aspect-ratio length radius ppi)
-    declare -a which_solved=(false false false false false false)
+    declare -a which_solved=(false false false false)
     if ! [ -z "$h" ]; then which_solved[0]=true; fi
     if ! [ -z "$w" ]; then which_solved[1]=true; fi
-    if ! [ -z "$ar" ]; then
-        regex="([0-9.]+):([0-9.]+)"
-        [[ $ar =~ $regex ]]
-        which_solved[2]=true;
-        ar=$(bc <<<"scale=2;${BASH_REMATCH[1]}/${BASH_REMATCH[2]}")
+    if ! [ -z "$r" ]; then which_solved[2]=true; fi
+    if ! [ -z "$ppi" ]; then which_solved[3]=true; fi
+    if [ -n "$h" -a -n "$w" -a "$w" -ne $((2 * h)) ]; then
+        echo "$bold${yellow}WARNING:$reset provided aspect ratio is not 2:1, recalculating h"
+        unset h
+        which_solved[0]=false
     fi
-    if ! [ -z "$l" ]; then which_solved[3]=true; fi
-    if ! [ -z "$r" ]; then which_solved[4]=true; fi
-    if ! [ -z "$ppi" ]; then which_solved[5]=true; fi
-    if [ -n "$h" -a -n "$w" -a -n "$ar" ] || [ -n "$r" -a -n "$l" -a -n "$ar" ]; then
-        echo "$bold${yellow}WARNING:$reset ignoring ar input and re-calculating"
-        unset ar
-        which_solved[2]=false
-    fi
-    if [ -n "$h" -a -n "$l" -a -n "$ppi" ] || [ -n "$w" -a -n "$r" -a -n "$ppi" ]; then
+    if [ -n "$w" -a -n "$r" -a -n "$ppi" ]; then
         echo "$bold${yellow}WARNING:$reset ignoring ppi input and re-calculating"
         unset ppi
-        which_solved[5]=false
+        which_solved[3]=false
     fi
-    if [ -n "$h" ] && [ -n "$w" ] && [ -n "$r" ] && [ -n "$l" ]; then
-        echo "$bold${yellow}WARNING:$reset too many constraints, ignoring "
-        if ! [ -z "$ppi" ]; then echo -n "radius"; unset r; fi
-        if ! [ -z "$ar" ]; then echo -n "height"; unset h; fi
-        echo -n " and length inputs"
-        unset l
+    if [ -n "$h" ] && [ -n "$w" ] && [ -n "$r" ] && [ -n "$ppi" ]; then
+        echo "$bold${yellow}WARNING:$reset too many constraints, ignoring height and width inputs"
+        unset h
+        unset w
+        which_solved[0]=false
+        which_solved[1]=false
     fi
     # while slots are unsolved, keep looping and try to calculate any variables we have the prerequisites to figure out
     # if it loops without having solved any more variables, it's stuck due to insufficient input
-    while [ $(solved $which_solved) -ne 6 ]; do
+    while [ $(solved $which_solved) -ne 4 ]; do
         prev_solved=$(solved $which_solved)
         if [ -z "$h" ]; then
             if [ -n "$w" ] && [ -n "$ar" ]; then
@@ -246,11 +213,14 @@ if [[ $analyze -eq 1 ]]; then
             fi
         fi
     done
-    summary "$w" "$h" "$ppi" "$thickness"
+    summary "$w" "$h" "$ppi" "$depth"
     exit 0
 fi
 # error if image isn't specified
-# XXX
+if ! [ -f $img ]; then
+    echo "$bold${red}ERROR:$reset could not find input image '$img'"
+    exit 125
+fi
 # command changes depending on location of openscad binary, i.e. mac vs. linux
 unameOut="$(uname -s)"
 case "${unameOut}" in
@@ -278,10 +248,15 @@ if ! [[ $h -gt 0 && $w -gt 0 ]] ; then
         h=${BASH_REMATCH[2]}
     fi
 fi
+# error if aspect ratio isn't 2:1
+if [ "$w" -ne $((2 * h)) ]; then
+    echo "$bold${red}ERROR:$reset input image or resize dimensions must have a 2:1 aspect ratio"
+    exit 125
+fi
 # convert image to texture, generate openscad file, and use openscad to generate stl file for 3d printing
 temp_dir=$(mktemp -d)
 python3 ~/Documents/OpenSCAD/libraries/BOSL2/scripts/img2scad.py "$img" -o "$temp_dir"/texture.scad -v image_array -r "${w}"x"${h}" > /dev/null 2>&1 &
-printf "include <BOSL2/std.scad>\ninclude <$temp_dir/texture.scad>\n\n\$fn= \$preview ? 20 : 200;\n\n//desired ppi has to be specified so the thickness and embossing depth can be consistent\nppi = $ppi;\nh = $h;\nw = $w;\n\nradius = w/(2*PI);\nheight = h;\npath = circle(r=radius);\nutomm = ppi/25.4;\n\n// how much the texture extrudes in mm (3-8)\ndepth = $depth * utomm;\n// how thick the cylinder is (5mm)\nthickness = $thickness * utomm;\n\ndifference() {\n    linear_sweep(\n        path, texture=image_array, tex_inset=false, tex_depth=-depth, tex_size=[h, w],\n        h=height, style=\"alt\");\n    down(10) cylinder(h=height + 20, r=radius - thickness);\n}" > "$temp_dir/roller.scad"
+printf "include <$temp_dir/texture.scad>\n\$fn= \$preview ? 20 : 200;\n\n//desired ppi has to be specified so the thickness and embossing depth can be consistent\nppi = $ppi;\nh = $h;\nw = $w;\nr = w/(2 * PI);\nutomm = ppi/25.4;\n// how much the texture extrudes in mm (3-8)\ndepth = $depth * utomm;\n\npoints = [ for (i = [ 0 : h ], j = [ 0 : w - 1 ]) let (radius = i < h ? r + depth * (1 - leapord[i][j]) : r + depth, theta = toDegrees(PI * 2 * j / w), phi = toDegrees(PI * i / h)) [radius * cos(theta) * sin(phi), radius * sin(theta) * sin(phi), radius * cos(phi)] ];\n\nfaces = [ for (i = [ 0 : h - 1 ], j = [ 0 : w - 1 ]) let (p1 = i * w + j, p3 = (i + 1) * w + (j + 1)) each [[p1, p3, (i + 1) * w + j], [p1, i * w + (j + 1), p3]]];\n\npolyhedron(points = points, faces = faces, convexity = 10);" > "$temp_dir/roller.scad"
 $cmd -q --export-format stl -o "$dir"/"$output" --backend Manifold "$temp_dir"/roller.scad & pid=$!
 # spinner logic, remove this and the & pid=$! on the line above to remove spinner
 trap 'kill $pid; exit' INT
